requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
pydantic==2.11.1
email-validator==2.1.1
python-dotenv==1.0.1
pymysql==1.1.0
